refactor(runtime): remove nvk_rt.h includes and update extern declarations
Removed unnecessary includes of nvk_rt.h from various header files. Updated variable declarations to use 'extern' for shared variables in kallsyms.h, log.h, and process.h. This change improves code clarity and reduces dependencies across headers.

+++ runtime/android/kernel/include/nvk_timer.h
@@ -83,9 +83,9 @@ static __always_inline u64 neverc_krt_arch_counter(void)
 
 static __always_inline u32 neverc_krt_arch_counter_freq(void)
 {
-	u32 freq;
+	u64 freq;
 	__asm__ __volatile__("mrs %0, cntfrq_el0" : "=r"(freq));
-	return freq;
+	return (u32)freq;
 }
 
 static __always_inline u64 neverc_krt_arch_counter_to_ns(u64 ticks)
